refactor: Change inner queue type to VecDeque and optimize Sender and Reciever structs
The changes include switching the inner queue type from being wrapped in a Mutex to directly using VecDeque. The Sender struct has been updated to have a shared Arc of Shared instead of Inner. The Reciever struct is also updated to use a shared Arc of Shared. Additionally, a new implementation of Drop for Sender has been added to handle tracking and notifying when all senders are dropped. The recv method in Reciever has been changed to return an Option<T> instead of T. Lastly, new test functions have been added for send and recv functionality and handling of dropped senders.