refactor: Improve buffer handling in Reciver<T> struct
Previously, the Reciver<T> struct did not efficiently handle the buffer for incoming data. This commit refactor adds a buffer VecDeque<T> to improve handling of incoming data. Now, when receiving data, if the buffer is not empty, it will return data from the buffer first before getting data from the main queue. This optimizes the data retrieval process and improves overall performance.